## Summary
- Add `GET /api/v1/words` and `GET /api/v1/words/overview` for the learning list (words with progress only)
- Rename advanced search to `POST /api/v1/words/search`
- Retire QAW / captured-words inbox — new words enter the learning list directly via `saveNewActiveWord`
- Extend `WordListItem` with `createdAt` and `definition`; publish `@kacper-ksiazek/ord-api-types` **1.0.47**
- Update `project-brief.md` and related docs to document the QAW retirement

## Test plan
- [x] Run word CRUD/overview integration tests
- [ ] Verify `GET /words` returns learning-list words only
- [ ] Publish types package 1.0.47 before merging frontend PR

+++ src/main/kotlin/com/ord/features/conversation/repositories/impl/ConversationMessageRepositoryCustomMethodsImpl.kt
@@ -68,7 +68,7 @@ class ConversationMessageRepositoryCustomMethodsImpl(
             .map { row ->
                 DailyActivityCount(
                     date = row.get("activity_date", LocalDate::class.java)!!,
-                    count = row.get("cnt", java.lang.Long::class.java)!!.toLong(),
+                    count = row.get("cnt", Long::class.java)!!,
                 )
             }
             .all()
